+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- [**BC**] Add `$initiator` argument to `ResponseDecoderInterface::supports` method
 
 ## 1.1.0 - 2021-07-28
 - Copy `$response` in `ProfilerItem` not to store a reference to the original object